third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-ice.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>

async_test(function(t) {
    const localPeerConnection = new RTCPeerConnection();
    localPeerConnection.close();

    let counter = 0;
    window.events = [];

    Promise.resolve().then(_ => events[counter++] = 1);

    const iceCandidate = new RTCIceCandidate({candidate: 'nano nano'});

    const unexpectedCallback = t.step_func(function() {
        assert_unreached('Unexpected callback');
    });

    localPeerConnection.addIceCandidate(iceCandidate, unexpectedCallback, t.step_func(error => {
        assert_equals(error.name, 'InvalidStateError');
        assert_equals(error.toString(), 'InvalidStateError: The RTCPeerConnection\'s signalingState is \'closed\'.');
        events[counter++] = 2;
    }));

    Promise.resolve().then(t.step_func_done(_ => {
        events[counter++] = 3;
        assert_array_equals(events, [1, 2, 3]);
    }));
}, 'Tests closed connection execution order.');

async_test(function(t) {
    const unexpectedCallback = t.step_func(function() {
        assert_unreached('Unexpected callback');
    });

    const expectedTypeError = t.step_func(function(error) {
      assert_equals(error.name, 'TypeError');
      calledCallbacks.push('expectedTypeError');
    });

    const addIceCandidateSuccess1 = t.step_func(function() {
        calledCallbacks.push('success1');
        pc.addIceCandidate({candidate: 'candidate', sdpMid: 0}, addIceCandidateSuccess2, unexpectedCallback);
    });

    const addIceCandidateSuccess2 = t.step_func(function() {
        calledCallbacks.push('success2');
        pc.addIceCandidate({candidate: 'candidate', sdpMLineIndex: 0}, addIceCandidateSuccess3, unexpectedCallback);
    });

    const addIceCandidateSuccess3 = t.step_func(function() {
        calledCallbacks.push('success3');
        pc.addIceCandidate({candidate: 'candidate', sdpMid: 0, sdpMLineIndex: 0}, addIceCandidateSuccess4, unexpectedCallback);
    });

    const addIceCandidateSuccess4 = t.step_func(function() {
        calledCallbacks.push('success4');
        pc.oniceconnectionstatechange = t.step_func_done(onIceChange2);
        pc.close();
    });

    const onIceChange1 = t.step_func(function() {
        if (pc.iceConnectionState === 'completed') {
            calledCallbacks.push('iceConnectionState complete');
            const iceCandidate = new RTCIceCandidate({candidate: 'nano nano'});
            pc.addIceCandidate(null, unexpectedCallback, unexpectedCallback).catch(expectedTypeError);
            pc.addIceCandidate({candidate: 'candidate'}, unexpectedCallback, unexpectedCallback).catch(expectedTypeError);
            pc.addIceCandidate(iceCandidate, null, unexpectedCallback).catch(expectedTypeError);
            pc.addIceCandidate(iceCandidate, unexpectedCallback, null).catch(expectedTypeError);
            pc.addIceCandidate(iceCandidate, addIceCandidateSuccess1, unexpectedCallback);
        }
    });

    const onIceChange2 = t.step_func_done(function() {
        if (pc.iceConnectionState === 'closed') {
            calledCallbacks.push('iceConnectionState closed');
            assert_array_equals(calledCallbacks, [
              'iceConnectionState complete',
              'expectedTypeError',
              'expectedTypeError',
              'expectedTypeError',
              'expectedTypeError',
              'success1',
              'success2',
              'success3',
              'success4',
              'iceConnectionState closed'
            ]);
        }
    });

    const calledCallbacks = [];
    const pc = new RTCPeerConnection();
    pc.oniceconnectionstatechange = t.step_func(onIceChange1);
}, 'Tests the RTCPeerConnection Ice functionality.');
</script>
</body>
</html>
